requirements.txt
openpyxl>=3.1.0
pandas>=2.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
click>=8.1.0
flask>=3.0.0
markdown>=3.5.0
python-docx>=1.0.0
fpdf2>=2.7.0
PyYAML>=6.0
html2text>=2024.0
pdfplumber>=0.10.0
xmltodict>=0.13.0
sqlparse>=0.4.4
Pillow>=10.0.0
croniter>=1.4.1
python-dateutil>=2.8.2
flask-limiter>=3.5.0
flask-compress>=1.14
bleach>=6.1.0
